requirements.txt
django
git+https://github.com/pennersr/django-allauth.git#egg=django-allauth
django-extensions
django-debug-toolbar
werkzeug
django-jfu
south
